refactor: 컬렉션 파라미터 방어 로직 추가 및 프로필 조회 로직 개선

+++ src/main/java/me/splleat/messengerproject/domain/channel/ChannelUserSettingService.java
@@ -66,7 +66,9 @@ public class ChannelUserSettingService {
 
     @Transactional
     public void leaveChannels(long userId, List<Long> channelIds) {
-        channelUserSettingRepository.deleteAllByUserIdAndChannelIdIn(userId, channelIds);
+        if (!channelIds.isEmpty()) {
+            channelUserSettingRepository.deleteAllByUserIdAndChannelIdIn(userId, channelIds);
+        }
     }
 
     @Transactional
